ImportSourceSelectionView removed now the designs updated to move that UI to the hub screen
