--- iOS/DuckDuckGo/DataImport/Hub/ImportSourceSelectionViewController.swift
@@ -1,43 +1,0 @@
-//
-//  ImportSourceSelectionViewController.swift
-//  DuckDuckGo
-//
-//  Copyright © 2026 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import UIKit
-import SwiftUI
-
-final class ImportSourceSelectionViewController: UIViewController {
-
-    private let viewModel = ImportSourceSelectionViewModel()
-    var onSourceSelected: ((ImportPasswordSource) -> Void)?
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        title = UserText.importSourceSelectionTitle
-        setupView()
-    }
-
-    private func setupView() {
-        viewModel.onSourceSelected = { [weak self] source in
-            self?.onSourceSelected?(source)
-        }
-
-        let hostingController = UIHostingController(rootView: ImportSourceSelectionView(viewModel: viewModel))
-        hostingController.view.backgroundColor = .clear
-        installChildViewController(hostingController)
-    }
-}
